+++ src/quiver/setup/__init__.py
@@ -1,0 +1,5 @@
+"""Setup wizard package."""
+
+from quiver.setup.commands import cmd_harness, cmd_setup
+
+__all__ = ["cmd_setup", "cmd_harness"]